adding showWorkflowCancellationDialog to editTemplate or editScript; creating getContentModelSnippets to retrieve contentModelSnippets; updating CodeEditorDialogContainer.tsx to use getContentModelSnippets to show the contentModel snippets;
